fix dialogs, use appVisibility plugin
